Refactor how isModalOpen gets updated
Before it was tightly coupled with the handleSubmit function.
This caused a bug because the function stops once it finish validating
all coordinates. Because of this setIsModalOpen never got executed.
This commit places the fetchEndTime function inside of a useEffect.
Therefore, it is always listening for isWaldoFound, isOdlawFound, and
isWizardFound, which then updates the isModalOpen.